+++ engine/admin-api/adapter/service/user/revoke_product_grants.go
@@ -2,13 +2,10 @@ package user
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"slices"
 
-	"github.com/konstellation-io/kai/engine/admin-api/adapter/config"
 	"github.com/konstellation-io/kai/engine/admin-api/domain/service/auth"
-	"github.com/spf13/viper"
 )
 
 func (ur *KeycloakUserRegistry) RevokeProductGrants(ctx context.Context, userEmail, product string, grants []auth.Action) error {
@@ -26,22 +23,5 @@ func (ur *KeycloakUserRegistry) RevokeProductGrants(ctx context.Context, userEma
 		return slices.Contains(grants, e)
 	})
 
-	marshalledRoles, err := json.Marshal(userGrantsByProduct)
-	if err != nil {
-		return fmt.Errorf("marshaling user's product grants: %w", err)
-	}
-
-	(*user.Attributes)["product_roles"] = []string{string(marshalledRoles)}
-
-	err = ur.refreshToken(ctx)
-	if err != nil {
-		return err
-	}
-
-	err = ur.client.UpdateUser(ctx, ur.token.AccessToken, viper.GetString(config.KeycloakRealmKey), *user)
-	if err != nil {
-		return fmt.Errorf("updating user: %w", err)
-	}
-
-	return nil
+	return ur.updatedUserWithNewGrants(ctx, user, userGrantsByProduct)
 }
